refactor: migrate simulation coordinates to integer game units
Simulation positions and spatial distances (radii, ranges, offsets, pushback) are now integer game units at 1000 units per tile, centralized in GameUnits. The arena spans 18000 by 32000 units; grid indices, time, damage, angles and multipliers keep their units.

- Position stores whole units with a fixed-point carry so fractional per-tick movement keeps its effective speed; clamping only touches out-of-bounds axes
- Range, sight, radius and collision checks use exact long squared distances
- Loaders convert each JSON field once based on its current unit (tile fields via tiles(), raw speeds to units per second, raw pushback passed through); resource JSON is unchanged
- Formation offsets are converted once; the circular fallback keeps its layout, with the 355 divisor renamed LEGACY_SUMMON_RADIUS_DIVISOR
- Per-tick arrival checks allow Position.MAX_ROUNDING_DISTANCE so rounding does not delay dashes, hooks or projectile hits by a tick
- Python bridge protocol and desktop rendering/input stay in tiles through boundary adapters
- Tests migrated to the same gameplay situations, plus new tests for conversions, movement, boundaries, overflow, formations and interface conversions

+++ core/src/main/java/org/crforge/core/ability/AbilityComponent.java
@@ -33,8 +33,9 @@ public class AbilityComponent {
   private DashState dashState = DashState.IDLE;
   private float dashTimer = 0f;
   private float dashCooldownTimer = 0f;
-  private float dashTargetX = 0f;
-  private float dashTargetY = 0f;
+  // Dash landing point in game units and dash speed in game units per second
+  private int dashTargetX = 0;
+  private int dashTargetY = 0;
   private float dashSpeed = 0f;
   // True once a target entered the acquisition range [minRange, maxRange].
   // While acquired, the target stays valid as long as it is within maxRange.
@@ -49,10 +50,11 @@ public class AbilityComponent {
 
   // TUNNEL state (Miner underground travel)
   private TunnelState tunnelState = TunnelState.INACTIVE;
-  private float tunnelTargetX = 0f;
-  private float tunnelTargetY = 0f;
-  private float tunnelWaypointX = 0f;
-  private float tunnelWaypointY = 0f;
+  // Tunnel destination and river-crossing waypoint in game units
+  private int tunnelTargetX = 0;
+  private int tunnelTargetY = 0;
+  private int tunnelWaypointX = 0;
+  private int tunnelWaypointY = 0;
   private boolean tunnelUsingWaypoint = false;
 
   // STEALTH state (Royal Ghost)
