+++ core/src/main/java/org/crforge/core/ability/BuffAllyAbility.java
@@ -5,19 +5,20 @@ import java.util.List;
 /**
  * BUFF_ALLY ability data (Rune Giant / GiantBuffer). Every {@code cooldown} seconds (after an
  * initial {@code actionDelay}), targets the {@code maxTargets} closest friendly troops within
- * {@code searchRange} tiles and applies a damage buff. Buffed troops deal bonus damage on every
- * {@code attackAmount}th attack. The buff persists for {@code persistAfterDeath} seconds after the
- * source dies.
+ * {@code searchRange} game units and applies a damage buff. Buffed troops deal bonus damage on
+ * every {@code attackAmount}th attack. The buff persists for {@code persistAfterDeath} seconds
+ * after the source dies.
  *
  * @param addedDamage base bonus damage per proc (level-1, scaled at spawn)
  * @param addedCrownTowerDamage base bonus crown tower damage per proc (level-1, scaled at spawn)
  * @param attackAmount number of attacks between procs (e.g. 3 = every 3rd attack)
- * @param searchRange range in tiles to search for friendly targets
+ * @param searchRange range in game units to search for friendly targets
  * @param maxTargets maximum number of friendlies to buff per cycle
  * @param cooldown seconds between buff cycles
  * @param actionDelay initial delay before first buff cycle
  * @param buffDelay delay before buff becomes active on target (simulates projectile travel)
- * @param maxRange maximum range before buff is dropped (unused in sim, kept for data parity)
+ * @param maxRange maximum range in game units before buff is dropped (unused in sim, kept for data
+ *     parity)
  * @param persistAfterDeath seconds the buff persists after the source GiantBuffer dies
  * @param damageMultipliers per-projectile/unit damage multiplier overrides
  */
@@ -25,12 +26,12 @@ public record BuffAllyAbility(
     int addedDamage,
     int addedCrownTowerDamage,
     int attackAmount,
-    float searchRange,
+    int searchRange,
     int maxTargets,
     float cooldown,
     float actionDelay,
     float buffDelay,
-    float maxRange,
+    int maxRange,
     float persistAfterDeath,
     List<DamageMultiplierEntry> damageMultipliers)
     implements AbilityData {
